+++ src/src/BelarusianHen.java
@@ -1,12 +1,14 @@
-public class UkrainianHen extends Hen{
+package src;
+
+public class BelarusianHen extends Hen{
         @Override
         public int getCountOfEggsPerMonth(){
-            return 6;
+            return 8;
         }
         @Override
         public String getDescription(){
             return super.getDescription() + " Моя страна - " +
-            Country.UKRAINE + ". Я несу " + this.getCountOfEggsPerMonth() +
+            Country.BELARUS + ". Я несу " + this.getCountOfEggsPerMonth() +
             " яиц в месяц.";
         }
     }
