+++ src/src/SomeClass.java
@@ -1,0 +1,6 @@
+package src;
+public class SomeClass {
+    public static String getString1() {
+        return "Just some class";
+    }
+}
