+++ src/src/UkrainianHen.java
@@ -1,12 +1,14 @@
-public class RussianHen extends Hen{
+package src;
+
+public class UkrainianHen extends Hen{
         @Override
         public int getCountOfEggsPerMonth(){
-            return 5;
+            return 6;
         }
         @Override
         public String getDescription(){
             return super.getDescription() + " Моя страна - " +
-            Country.RUSSIA + ". Я несу " + this.getCountOfEggsPerMonth() +
+            Country.UKRAINE + ". Я несу " + this.getCountOfEggsPerMonth() +
             " яиц в месяц.";
         }
     }
